split view controller for person-list and person-details
